+++ include/Material.hpp
@@ -6,7 +6,7 @@
 
 class Material : public Loot {
     public:
-    Material(const std::string& nume, float greutate, int raritate);
+    Material(const std::string& nume, double greutate, int raritate);
     ~Material() override;
 };
