+++ src/Obiect.cpp
@@ -3,5 +3,5 @@
 using namespace std;
 
 
-Obiect::Obiect(const string &nume, float greutate) : nume(nume), greutate(greutate) {}
+Obiect::Obiect(const string &nume, double greutate) : nume(nume), greutate(greutate) {}
 Obiect::~Obiect() = default;
